+++ example/lib/main.dart
@@ -30,6 +30,9 @@ void main() async {
   FlutterBranchSdk.setRequestMetadata('key2', 'value2');
   */
 
+  FlutterBranchSdk.setAnonID('1234556');
+  FlutterBranchSdk.setSDKWaitTimeForThirdPartyAPIs(2.5);
+
   await FlutterBranchSdk.init(enableLogging: true, branchAttributionLevel: BranchAttributionLevel.FULL);
   FlutterBranchSdk.setConsumerProtectionAttributionLevel(BranchAttributionLevel.FULL);