+++ lib/src/flutter_branch_sdk.dart
@@ -246,4 +246,17 @@ class FlutterBranchSdk {
   static void setConsumerProtectionAttributionLevel(BranchAttributionLevel branchAttributionLevel) {
     FlutterBranchSdkPlatform.instance.setConsumerProtectionAttributionLevel(branchAttributionLevel);
   }
+
+  /// Sets a custom Meta Anon ID for the current user.
+  /// [anonID] The custom Meta Anon ID to be used by Branch.
+  static void setAnonID(String anonId) {
+    FlutterBranchSdkPlatform.instance.setAnonID(anonId);
+  }
+
+  /// Set the SDK wait time for third party APIs (for fetching ODM info and Apple Attribution Token) to finish
+  /// This timeout should be > 0 and <= 10 seconds.
+  /// [waitTime] Number of seconds before third party API calls are considered timed out. Default is 0.5 seconds (500ms).
+  static void setSDKWaitTimeForThirdPartyAPIs(double waitTime) {
+    FlutterBranchSdkPlatform.instance.setSDKWaitTimeForThirdPartyAPIs(waitTime);
+  }
 }