+++ lib/src/flutter_branch_sdk_platform_interface.dart
@@ -256,4 +256,17 @@ abstract class FlutterBranchSdkPlatform extends PlatformInterface {
   void setConsumerProtectionAttributionLevel(BranchAttributionLevel branchAttributionLevel) {
     throw UnimplementedError('setConsumerProtectionAttributionLevel has not been implemented');
   }
+
+  /// Sets a custom Meta Anon ID for the current user.
+  /// [anonID] The custom Meta Anon ID to be used by Branch.
+  void setAnonID(String anonId) {
+    throw UnimplementedError('setAnonID has not been implemented');
+  }
+
+  /// Set the SDK wait time for third party APIs (for fetching ODM info and Apple Attribution Token) to finish
+  /// This timeout should be > 0 and <= 10 seconds.
+  /// [waitTime] Number of seconds before third party API calls are considered timed out. Default is 0.5 seconds (500ms).
+  void setSDKWaitTimeForThirdPartyAPIs(double waitTime) {
+    throw UnimplementedError('setSDKWaitTimeForThirdPartyAPIs has not been implemented');
+  }
 }